ref(onboarding): Split fastapi onboarding docs (#103045)
Contributes to
https://linear.app/getsentry/issue/TET-864/introduce-folders-for-onboarding-platforms
