+++ static/app/gettingStartedDocs/python/fastapi/onboarding.spec.tsx
@@ -6,7 +6,7 @@ import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
 
-import docs from './fastapi';
+import docs from '.';
 
 describe('flask onboarding docs', () => {
   it('renders doc correctly', () => {
